requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
openai==1.51.2
python-dotenv==1.0.1
pydantic==2.12.2